Move code not related to Cloud out of CloudAwareProvisioner
This allows it to be used as a base class for other Provisioners that do
not necessarily have only a single VMType (and so cannot be used with an
instance of Environment).
